Enhance error messages logging the statement executed with error
Before this change the only information logged was the script filename.
This created a problem on a script with multiple statements, you had to
dig into the script to understand which statement was the incorrect one,
by splitting the script into statements and executing them one by one
makes it easier to debug.